requirements.txt
discord.py==2.2.2
python-dotent==1.0.0